# OpenAPI.yaml
openapi: 3.0.0
info:
  title: MY SWEET SKRIPSI
  contact: {}
  version: "1.0.0"
servers:
  - url: http://localhost:8000/v1/api
  - url: https://be-permission.up.railway.app/v1/api
    variables: {}
paths:
...
  /pegawai/register:
    post:
      tags:
        - For Data Pegawai
      summary: REGISTER PEGAWAI
      description: ""
      operationId: REGISTERPEGAWAI
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              type: object
              properties:
                name_pegawai:
                  type: string
                  description: "insert name"
                sex:
                  type: integer
                  description: "1 for men, and 0 for women"
                email:
                  type: string
                  format: email
                  description: "insert name"
                password:
                  type: string
                  description: "insert password"
                role_id:
                  type: integer
                  description: "choose your role, 1. admin, 2. ustadz, 3. security"
              required:
                - name_pegawai
                - sex
                - email
                - password
                - role_id
            examples:
              example1:
                value:
                  name_pegawai: "Hendri Edited"
                  sex: 1
                  email: "hendripermana60@gmail.com123"
                  password: "12345cekceksaja"
                  role_id: 2
      responses:
        "200":
          description: Successful create data pegawai
          headers: {}
          content: {}
        "400":
          description: Error or not success create data pegawai
          headers: {}
          content: {}
        "500":
          description: internal server error
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security: []
...